Removed the github mirror: git was causing too much hassle.
I had originally planned to maintain the mirror on github for as long as
possible, but github is refusing to let me push the latest changes.
I've spent several hours trying to fix this, but have given up.  I moved
to Mercurial after losing a lot of code in git (even being the only
person using the repository) and after fighting it for several months
and having repeated problems I moved the core repository over to
Mercurial; I'm now making that the only supported VCS.